Linking flag
This change sets the rpath variable in the iii binary so that,
if you have installed libz3.so at {path_to_executable}/../lib/libz3.so,
then that library will be dynamically loaded when you run iii.

This fixes a problem we have been having with the mk_install.py script
where it would give priority to the copy found in the opam directory
instead of using the copy created at the time that we made the install.

To see the impact of this change, run this command

    $ readelf -d _build/install/default/bin/asli | grep R*PATH
     0x000000000000001d (RUNPATH)            Library runpath: [$ORIGIN/../lib:/home/areid/.opam/5.3.0/lib/stublibs]

This shows that the searchpath for dynamic libraries has
$ORIGIN/../lib as the first (highest priority) directory
with the opam directory as a fallback in case that fails.

So, if you are using the result of the mk_install.py script, it will use
the libz3.so copy made by that script.

And, if you are using 'dune build', it will use the copy in your
opam directory.