Create specialized versions of function which has prototype only
Useful if the function is imported, using 'foreign import' declaration.
In this case it does not have/need the definition in ISA.

Previously the monomorphization processed function definitions only and
created corresponding function prototype for each specialized version of
a function definition at the final stage.

Now it creates specialized versions of function prototype and function
definition simultaneously.

  * The function lookup (hash) table, with function identifier as the
    key, now can contain both function prototype and definition (if
    available) as value.

  * Since function definition is accompanied by prototype, the
    implicit/explicit requests are based on prototypes only for
    simplicity.

  * The map of instances now has list of monomorphic declarations as
    value, instead of one declaration. The list can contain both
    function prototype and definition (if available).