Modify handling of --split-state to support multithreading
If you specify 'should_add_fun_arg' in the split_state part
of a configuration file, then an extra argument is added to
every function that points to a struct containing that
part of the state.

This extra argument is also added to FFI exported functions:
the client system is expected to allocate, initialize and
pass a pointer to the state using this extra argument.

This extra argument is also added to FFI imported functions:
this allows the imported function to call an FFI exported function
on the same thread or to access any client fields that have
been added to the struct.

A typical entry in the split_state part of a configuration
file looks like this

        "threadlocal_state" : {
            "variables": [".*"],
            "client_fields": ["void* client_ptr;"],
            "should_add_fun_arg": ["yes"]
        }

If the 'should_add_fun_arg' is '["yes"]' then this part
of the state will be accessed via a function argument as
described above.
Otherwise, it will be accessed via a global variable that
is expected to point to this part of the state.