Discard code not reachable from exported functions later
Since exporting specialized versions of functions using 'foreign export
function' declarations is supported, discarding the code not reachable
from the functions at the first transformation is not possible because
the specialized versions are generated after the monomorphize phase
only.

One option to solve this is to avoid removing unreachable code until the
monomorphization is completed.

Other option is to take two step approach: (1) delete code not reachable
from the corresponding polymorphic functions, and (2) later delete code
not reachable from the specialized versions of the functions.

Implement the 1st option for now as it looks simple and clean.