AC/SC runtime optimization: compensate for the 'nearest_intsize' optimization
The nearest_intsize optimization was causing us to generate code like

    int8_t x;
    ....
    if (int8_t)x < ... then

where the cast on the last line clearly pointless because x already has
type int8_t.

The reason that we were generating code like this is that x
might have ISA type '__sint(5)' but the use of x on the
final line requires us to change it to ISA type '__sint(6)'.
Since 5 != 6, we insert a cast to change the type.
But, thanks to the nearest_intsize optimization, both ISA
types were being represented by the C type int8_t so the
cast is not necessary.

Note that this change is not strictly speaking necessary because
the C compiler is quite happy to cast an int8_t value to int8_t.
But, the extra cast adds a bit of clutter to the generated code
that makes it a little harder for humans to read.