xform_bounded: de-optimize for loop variables
The :xform_bounded transformation was being a bit cute in how it figured
out the bounds on the loop control variable in a for loop.

In a loop like 'for i := 0 to 15 do ... endfor;',
we will use 'i' in two ways.

1. We use 'i' as a counter to track whether we are at the end of
   the loop. For this purpose, i needs to have the range {0..16}
   because, on the final iteration, i will be incremented past the
   end of the range.

2. We probably use 'i' inside the loop - perhaps as an array index (say).
   Inside the loop, i can only have the range {0..15} because we will
   not execute the loop body with any other value.

The problem with trying to exploit this is that the representation of
'i' has to be at least 5 bits even though a range of {0..15} might
suggest that it would only be 4 bits.

This change makes us use the same bounds in both places that 'i' is
used - avoiding the type mismatch that we were seeing.