C backend: optimize 32-bit constants
When a 32-bit constant is used as an index, we can use the
constant directly without having to convert to some other type and
then back to uint64_t.

(It is probably safe to support 63 or 64 bit constants here but
indexes that large will be causing a bunch of other problems
like overflowing memory, etc. so let's stay well clear of that
kind of value.)