Demo/test monomorphization of another recursion pattern
This test (recurse_02.isa) differs from the previous one (recurse_01.isa) in that
the recursion always respects power-of-two alignment boundaries.
That is, it is fine to add x0 and x1 because they are in the same block
but it is not ok to add x1 and x2 because they are in different blocks.

This is intended to match the behavior of vector tree reduction operations.