Demo: don't #include .c files
We were using a dirty trick of #including .c files instead
of compiling them separately.
This is sort of ok (ish) on small codebases but is not how you would
do things on larger projects.

Changing it to compile the .c files separately to make it more like the
way that larger projects are built - which makes this a better test
for whether we are generating correct code.